# target/generated-sources/openapi/src/main/java/com/portfolio/catalog/generated/api/ProductsApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.5.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.portfolio.catalog.generated.api;

import com.portfolio.catalog.generated.model.CreateProductRequest;
import com.portfolio.catalog.generated.model.Product;
import com.portfolio.catalog.generated.model.ProductAvailabilityRequest;
import com.portfolio.catalog.generated.model.ProductPage;
import com.portfolio.catalog.generated.model.ProductSummary;
import java.util.UUID;
import com.portfolio.catalog.generated.model.UpdateProductRequest;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2025-11-06T08:24:45.086812400+01:00[Europe/Madrid]", comments = "Generator version: 7.5.0")
@Validated
@Tag(name = "Products", description = "Product catalog management")
public interface ProductsApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    /**
     * POST /products : Create product
     *
     * @param createProductRequest  (required)
     * @return Product created. (status code 201)
     *         or Product with same SKU already exists. (status code 409)
     */
    @Operation(
        operationId = "createProduct",
        summary = "Create product",
        tags = { "Products" },
        responses = {
            @ApiResponse(responseCode = "201", description = "Product created.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Product.class))
            }),
            @ApiResponse(responseCode = "409", description = "Product with same SKU already exists.")
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/products",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    default ResponseEntity<Product> createProduct(
        @Parameter(name = "CreateProductRequest", description = "", required = true) @Valid @RequestBody CreateProductRequest createProductRequest
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"lastUpdatedAt\" : \"2000-01-23T04:56:07.000+00:00\", \"price\" : 0.08008281904610115, \"name\" : \"name\", \"description\" : \"description\", \"stockQuantity\" : 0, \"currency\" : \"EUR\", \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\", \"sku\" : \"sku\", \"status\" : \"ACTIVE\", \"tags\" : [ \"tags\", \"tags\" ] }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * DELETE /products/{id} : Delete product
     *
     * @param id Unique product identifier. (required)
     * @return Product deleted successfully. (status code 204)
     */
    @Operation(
        operationId = "deleteProduct",
        summary = "Delete product",
        tags = { "Products" },
        responses = {
            @ApiResponse(responseCode = "204", description = "Product deleted successfully.")
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = "/products/{id}"
    )
    
    default ResponseEntity<Void> deleteProduct(
        @Parameter(name = "id", description = "Unique product identifier.", required = true, in = ParameterIn.PATH) @PathVariable("id") UUID id
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /products/{id} : Get product by id
     *
     * @param id Unique product identifier. (required)
     * @return Product found. (status code 200)
     *         or Product missing. (status code 404)
     */
    @Operation(
        operationId = "getProduct",
        summary = "Get product by id",
        tags = { "Products" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Product found.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Product.class))
            }),
            @ApiResponse(responseCode = "404", description = "Product missing.")
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/products/{id}",
        produces = { "application/json" }
    )
    
    default ResponseEntity<Product> getProduct(
        @Parameter(name = "id", description = "Unique product identifier.", required = true, in = ParameterIn.PATH) @PathVariable("id") UUID id
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"lastUpdatedAt\" : \"2000-01-23T04:56:07.000+00:00\", \"price\" : 0.08008281904610115, \"name\" : \"name\", \"description\" : \"description\", \"stockQuantity\" : 0, \"currency\" : \"EUR\", \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\", \"sku\" : \"sku\", \"status\" : \"ACTIVE\", \"tags\" : [ \"tags\", \"tags\" ] }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /products : List products
     * Returns a paginated list of products ordered by last update.
     *
     * @param page Page number (0-indexed). (optional)
     * @param size Page size. (optional)
     * @param sort Sort criteria, e.g. &#x60;name,asc&#x60;. (optional)
     * @return Page of products. (status code 200)
     */
    @Operation(
        operationId = "listProducts",
        summary = "List products",
        description = "Returns a paginated list of products ordered by last update.",
        tags = { "Products" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Page of products.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ProductPage.class))
            })
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/products",
        produces = { "application/json" }
    )
    
    default ResponseEntity<ProductPage> listProducts(
        @Min(0) @Parameter(name = "page", description = "Page number (0-indexed).", in = ParameterIn.QUERY) @Valid @RequestParam(value = "page", required = false) Integer page,
        @Min(1) @Max(100) @Parameter(name = "size", description = "Page size.", in = ParameterIn.QUERY) @Valid @RequestParam(value = "size", required = false) Integer size,
        @Parameter(name = "sort", description = "Sort criteria, e.g. `name,asc`.", in = ParameterIn.QUERY) @Valid @RequestParam(value = "sort", required = false) String sort
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"size\" : 5, \"totalPages\" : 2, \"page\" : 1, \"content\" : [ { \"lastUpdatedAt\" : \"2000-01-23T04:56:07.000+00:00\", \"price\" : 0.08008281904610115, \"name\" : \"name\", \"description\" : \"description\", \"stockQuantity\" : 0, \"currency\" : \"EUR\", \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\", \"sku\" : \"sku\", \"status\" : \"ACTIVE\", \"tags\" : [ \"tags\", \"tags\" ] }, { \"lastUpdatedAt\" : \"2000-01-23T04:56:07.000+00:00\", \"price\" : 0.08008281904610115, \"name\" : \"name\", \"description\" : \"description\", \"stockQuantity\" : 0, \"currency\" : \"EUR\", \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\", \"sku\" : \"sku\", \"status\" : \"ACTIVE\", \"tags\" : [ \"tags\", \"tags\" ] } ], \"totalElements\" : 5 }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * PATCH /products/{id}/availability : Toggle product availability
     *
     * @param id Unique product identifier. (required)
     * @param productAvailabilityRequest  (required)
     * @return Product availability updated. (status code 200)
     *         or Product missing. (status code 404)
     */
    @Operation(
        operationId = "patchAvailability",
        summary = "Toggle product availability",
        tags = { "Products" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Product availability updated.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Product.class))
            }),
            @ApiResponse(responseCode = "404", description = "Product missing.")
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.PATCH,
        value = "/products/{id}/availability",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    default ResponseEntity<Product> patchAvailability(
        @Parameter(name = "id", description = "Unique product identifier.", required = true, in = ParameterIn.PATH) @PathVariable("id") UUID id,
        @Parameter(name = "ProductAvailabilityRequest", description = "", required = true) @Valid @RequestBody ProductAvailabilityRequest productAvailabilityRequest
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"lastUpdatedAt\" : \"2000-01-23T04:56:07.000+00:00\", \"price\" : 0.08008281904610115, \"name\" : \"name\", \"description\" : \"description\", \"stockQuantity\" : 0, \"currency\" : \"EUR\", \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\", \"sku\" : \"sku\", \"status\" : \"ACTIVE\", \"tags\" : [ \"tags\", \"tags\" ] }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /products/search : Search products by text
     *
     * @param q Search query applied over name and description. (required)
     * @param limit Max number of results. (optional)
     * @return Matching search results. (status code 200)
     */
    @Operation(
        operationId = "searchProducts",
        summary = "Search products by text",
        tags = { "Products" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Matching search results.", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = ProductSummary.class)))
            })
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/products/search",
        produces = { "application/json" }
    )
    
    default ResponseEntity<List<ProductSummary>> searchProducts(
        @NotNull @Size(min = 2) @Parameter(name = "q", description = "Search query applied over name and description.", required = true, in = ParameterIn.QUERY) @Valid @RequestParam(value = "q", required = true) String q,
        @Min(1) @Max(50) @Parameter(name = "limit", description = "Max number of results.", in = ParameterIn.QUERY) @Valid @RequestParam(value = "limit", required = false) Integer limit
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "[ { \"name\" : \"name\", \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\", \"sku\" : \"sku\", \"status\" : \"status\" }, { \"name\" : \"name\", \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\", \"sku\" : \"sku\", \"status\" : \"status\" } ]";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * PUT /products/{id} : Update product by id
     *
     * @param id Unique product identifier. (required)
     * @param updateProductRequest  (required)
     * @return Product updated. (status code 200)
     *         or Product missing. (status code 404)
     */
    @Operation(
        operationId = "updateProduct",
        summary = "Update product by id",
        tags = { "Products" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Product updated.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Product.class))
            }),
            @ApiResponse(responseCode = "404", description = "Product missing.")
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.PUT,
        value = "/products/{id}",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    default ResponseEntity<Product> updateProduct(
        @Parameter(name = "id", description = "Unique product identifier.", required = true, in = ParameterIn.PATH) @PathVariable("id") UUID id,
        @Parameter(name = "UpdateProductRequest", description = "", required = true) @Valid @RequestBody UpdateProductRequest updateProductRequest
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"lastUpdatedAt\" : \"2000-01-23T04:56:07.000+00:00\", \"price\" : 0.08008281904610115, \"name\" : \"name\", \"description\" : \"description\", \"stockQuantity\" : 0, \"currency\" : \"EUR\", \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\", \"sku\" : \"sku\", \"status\" : \"ACTIVE\", \"tags\" : [ \"tags\", \"tags\" ] }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}